eg: add click delay, fixes #75
Playwright triggered the click such that the purchase frame opened, but
did not resolve the promise. Had to move mouse into the browser for it
to continue. Adding a click delay of 1ms also worked (default is no
delay between mouse down and up). Went for 11ms. A typical click is
probably 100-200ms.